requirements.txt
Django==5.0.4
django-phonenumber-field==7.3.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-phonenumbers==1.0.1
pillow==10.3.0
